Revert "save"
This reverts commit dd7a5787914aa8b94a9346a550866094de8e2822.

+++ SimRTS/Source/RTSEngine/Public/Types.h
@@ -6,7 +6,7 @@ namespace SimRTS {
 
 using Tick = int32_t;
 using UnitId = int32_t;
-using PlayerId = uint8_t;
+using PlayerId = int32_t;
 
 struct Vec2i {
     int32_t x = 0;
