+++ SimRTS/Source/SimRTS/Bridge/SimBridge.h
@@ -17,13 +17,13 @@ public:
 		int32 TargetX,
 		int32 TargetY,
 		bool bIsNext = false,
-		uint8 PlayerId = 0);
+		int32 PlayerId = 0);
 	void SubmitScheduledMoveOrder(
 		const TArray<int32>& UnitIds,
 		int32 TargetX,
 		int32 TargetY,
 		bool bIsNext,
-		uint8 PlayerId,
+		int32 PlayerId,
 		uint32 OrderId,
 		int32 ScheduledTick);
 	void StepForward();
